feat: Update Quiz logic and UI
- Modified `Question` data class to include `media`, `explanation`, `points`, `type`, `difficulty`, and `pointsGained`.
- Updated `Quiz` data class to store `quizTimer` and `answerLockTimer` as Long.
- Refactored `QuizTimerManager` to use Long for time-related values and updated time-related functions.
- Updated `QuizProgressIndicators` to accept Long values for time.
- Modified `formatTime` in `NumberExtensions` to accept Long for time formatting.
- Changed `totalQuizTime` in `QuizScreen` to Long.
- Added explanation to the `PointsAnimationScreen` composable and show it only if the answer is correct.
- fixed issue that the question was locked after answering it.

+++ app/src/main/java/com/netanel/xplore/quiz/model/Quiz.kt
@@ -6,8 +6,8 @@ import kotlinx.serialization.Serializable
 data class Quiz(
     var _id: String = "",
     var questions: List<Question>,
-    val quizTimer: Int,
-    val answerLockTimer: Int,
+    val quizTimer: Long,
+    val answerLockTimer: Long,
     val title: String? = null,
     val creatorId: String? = null,
     val isActive: Boolean = true,
